rename histogram renderer files
