+++ plugins/infovis/src/TableHistogramRenderer2D.h
